chore: remove re-export stubs from Phase 5 refactoring
Update imports to use direct paths instead of barrel re-exports:
- Grid components: @/features/grid-editor/components/Grid
- Shared hooks (useResponsive, etc.): @/shared/hooks
- Grid hooks (useGridCoords, etc.): @/features/grid-editor/hooks
- Domain hooks: @/hooks/useFeatureFlag, @/hooks/useCollabMode, etc.

Delete src/components/Grid/index.ts (no longer needed).
Simplify src/hooks/index.ts to only export hooks that live there.
Update Header.test.tsx mocks for new import paths.
