+++ src/features/grid-editor/components/Grid/index.tsx
@@ -6,7 +6,8 @@ import { useInteractionStore } from '@/core/store/interaction';
 import { useSelectionStore } from '@/core/store/selection';
 import { useHalfBinModeStore } from '@/core/store/halfBinMode';
 import { useToastStore } from '@/core/store/toast';
-import { useInteraction, useResponsive, useGridResize } from '@/hooks';
+import { useInteraction, useGridResize } from '@/features/grid-editor/hooks';
+import { useResponsive } from '@/shared/hooks';
 import { BASE_CELL_SIZE, STAGING_ID, CONSTRAINTS, getBaseCellSize, HALF_BIN_SCALE } from '@/core/constants';
 import { lazyWithRetry, namedExport } from '@/utils/lazyWithRetry';
 import { track3DPreview } from '@/utils/analytics';
