+++ src/test/components/Header.test.tsx
@@ -14,15 +14,22 @@ vi.mock('../../features/layout-library/components/LayoutManagerModal', () => ({
 // Controllable mock for useFeatureFlag and useCollabMode
 let mockFeatureFlagValue = false;
 let mockIsCollaborative = false;
-vi.mock('../../hooks', async (importOriginal) => {
+
+vi.mock('@/shared/hooks', async (importOriginal) => {
   const actual = await importOriginal();
   return {
     ...actual,
     useResponsive: () => ({ isTablet: false, isMobile: false }),
-    useFeatureFlag: () => mockFeatureFlagValue,
-    useCollabMode: () => ({ isCollaborative: mockIsCollaborative, canEdit: true, shareId: null }),
   };
 });
+
+vi.mock('@/hooks/useFeatureFlag', () => ({
+  useFeatureFlag: () => mockFeatureFlagValue,
+}));
+
+vi.mock('@/hooks/useCollabMode', () => ({
+  useCollabMode: () => ({ isCollaborative: mockIsCollaborative, canEdit: true, shareId: null }),
+}));
 
 // Controllable mock for ShareButton
 let mockShareButtonEnabled = false;
